OSCO result + Tanium result fixes.
