+++ docs/api_reference/trestle.tasks.osco_result_to_oscal_ar.md
@@ -1,0 +1,2 @@
+::: trestle.tasks.osco_result_to_oscal_ar
+handler: python
