--- docs/api_reference/trestle.tasks.osco_to_oscal.md
@@ -1,2 +1,0 @@
-::: trestle.tasks.osco_to_oscal
-handler: python
