--- docs/api_reference/trestle.tasks.tanium_report_to_oscal_ar.md
@@ -1,2 +1,0 @@
-::: trestle.tasks.tanium_report_to_oscal_ar
-handler: python
