+++ docs/api_reference/trestle.tasks.tanium_result_to_oscal_ar.md
@@ -1,0 +1,2 @@
+::: trestle.tasks.tanium_result_to_oscal_ar
+handler: python
