+++ docs/tutorials/task.tanium-result-to-oscal-ar/800px-Australia_scarsdale_nimons_bridge.copyright-info.txt
