+++ mkdocs.yml
@@ -26,7 +26,7 @@ nav:
   - Transformer construction: tutorials/task.transformer-construction/transformer-construction.md
   - Task - ocp4-cis-profile-to-oscal-catalog: tutorials/task.ocp4-cis-profile-to-oscal-catalog/transformation.md
   - Task - ocp4-cis-profile-to-oscal-cd: tutorials/task.ocp4-cis-profile-to-oscal-cd/transformation.md
-  - Task - tanium-report-to-oscal-ar: tutorials/task.tanium-report-to-oscal-ar/transformation.md
+  - Task - tanium-result-to-oscal-ar: tutorials/task.tanium-result-to-oscal-ar/transformation.md
   - Work with authoring versions: tutorials/work_with_authoring_versions.md
 - Trestle command-line interface (CLI):
   - CLI for OSCAL documents: cli.md
@@ -146,8 +146,8 @@ nav:
       - ocp4_cis_profile_to_oscal_catalog: api_reference/trestle.tasks.ocp4_cis_profile_to_oscal_catalog.md
       - ocp4_cis_profile_to_oscal_cd: api_reference/trestle.tasks.ocp4_cis_profile_to_oscal_cd.md
       - oscal_profile_to_osco_profile: api_reference/trestle.tasks.oscal_profile_to_osco_profile.md
-      - osco_to_oscal: api_reference/trestle.tasks.osco_to_oscal.md
-      - tanium_report_to_oscal_ar: api_reference/trestle.tasks.tanium_report_to_oscal_ar.md
+      - osco_result_to_oscal_ar: api_reference/trestle.tasks.osco_result_to_oscal_ar.md
+      - tanium_result_to_oscal_ar: api_reference/trestle.tasks.tanium_result_to_oscal_ar.md
       - transform: api_reference/trestle.tasks.transform.md
       - xlsx_helper: api_reference/trestle.tasks.xlsx_helper.md
       - xlsx_to_oscal_cd: api_reference/trestle.tasks.xlsx_to_oscal_cd.md
