+++ tests/data/tasks/osco/test-osco-result-to-oscal-ar-1.3.5.config
@@ -1,4 +1,4 @@
-[task.osco-to-oscal]
+[task.osco-result-to-oscal-ar]
 
 input-dir =  tests/data/tasks/osco/input-1.3.5
 output-dir = tests/data/tasks/osco/output-1.3.5
