+++ tests/data/tasks/osco/test-osco-result-to-oscal-ar-configmaps.config
@@ -1,4 +1,4 @@
-[task.osco-to-oscal]
+[task.osco-result-to-oscal-ar]
 
 input-dir =  tests/data/tasks/osco/input-configmaps
 output-dir = tests/data/tasks/osco/output-configmaps
