+++ tests/data/tasks/osco/test-osco-result-to-oscal-ar-xml-rhel7.config
@@ -1,4 +1,4 @@
-[task.osco-to-oscal]
+[task.osco-result-to-oscal-ar]
 
 input-dir =  tests/data/tasks/osco/input-xml-rhel7
 output-dir = tests/data/tasks/osco/output-xml-rhel7
