+++ tests/data/tasks/osco/test-osco-result-to-oscal-ar.config
@@ -1,4 +1,4 @@
-[task.osco-to-oscal]
+[task.osco-result-to-oscal-ar]
 
 input-dir =  tests/data/tasks/osco/input
 output-dir = tests/data/tasks/osco/output
