+++ tests/data/tasks/tanium/demo-tanium-result-to-oscal-ar.config
@@ -1,4 +1,4 @@
-[task.tanium-report-to-oscal-ar]
+[task.tanium-result-to-oscal-ar]
 
 input-dir =  tests/data/tasks/tanium/input
 output-dir = tests/data/tasks/tanium/runtime
