+++ tests/trestle/tasks/osco_result_to_oscal_ar_test.py
@@ -23,7 +23,7 @@ from _pytest.monkeypatch import MonkeyPatch
 
 from tests.test_utils import text_files_equal
 
-import trestle.tasks.osco_to_oscal as osco_to_oscal
+import trestle.tasks.osco_result_to_oscal_ar as osco_result_to_oscal_ar
 import trestle.transforms.implementations.osco as osco
 from trestle.tasks.base_task import TaskOutcome
 
@@ -43,11 +43,11 @@ class MonkeyBusiness():
 def test_osco_print_info(tmp_path):
     """Test print_info call."""
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     section['output-dir'] = str(tmp_path)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.print_info()
     assert retval is None
 
@@ -55,11 +55,11 @@ def test_osco_print_info(tmp_path):
 def test_osco_simulate(tmp_path):
     """Test simulate call."""
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     section['output-dir'] = str(tmp_path)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.simulate()
     assert retval == TaskOutcome.SIM_SUCCESS
     assert len(os.listdir(str(tmp_path))) == 0
@@ -68,11 +68,11 @@ def test_osco_simulate(tmp_path):
 def test_osco_simulate_compressed(tmp_path):
     """Test simulate call with compressed OSCO xml data."""
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal-compressed.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar-compressed.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     section['output-dir'] = str(tmp_path)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.simulate()
     assert retval == TaskOutcome.SIM_SUCCESS
     assert len(os.listdir(str(tmp_path))) == 0
@@ -80,7 +80,7 @@ def test_osco_simulate_compressed(tmp_path):
 
 def test_osco_simulate_no_config(tmp_path):
     """Test simulate no config call."""
-    tgt = osco_to_oscal.OscoToOscal(None)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(None)
     retval = tgt.simulate()
     assert retval == TaskOutcome.SIM_FAILURE
     assert len(os.listdir(str(tmp_path))) == 0
@@ -89,17 +89,17 @@ def test_osco_simulate_no_config(tmp_path):
 def test_osco_simulate_no_overwrite(tmp_path):
     """Test simulate no overwrite call."""
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     section['output-dir'] = str(tmp_path)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.SUCCESS
     assert len(os.listdir(str(tmp_path))) == 1
     section['output-overwrite'] = 'false'
     section['output-dir'] = str(tmp_path)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.simulate()
     assert retval == TaskOutcome.SIM_FAILURE
     assert len(os.listdir(str(tmp_path))) == 1
@@ -108,12 +108,12 @@ def test_osco_simulate_no_overwrite(tmp_path):
 def test_osco_simulate_no_input_dir(tmp_path):
     """Test simulate with no input dir call."""
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar.config')
     config.read(config_path)
-    config.remove_option('task.osco-to-oscal', 'input-dir')
-    section = config['task.osco-to-oscal']
+    config.remove_option('task.osco-result-to-oscal-ar', 'input-dir')
+    section = config['task.osco-result-to-oscal-ar']
     section['output-dir'] = str(tmp_path)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.simulate()
     assert retval == TaskOutcome.SIM_FAILURE
     assert len(os.listdir(str(tmp_path))) == 0
@@ -122,11 +122,11 @@ def test_osco_simulate_no_input_dir(tmp_path):
 def test_osco_simulate_no_ouput_dir(tmp_path):
     """Test simulate with no output dir call."""
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar.config')
     config.read(config_path)
-    config.remove_option('task.osco-to-oscal', 'output-dir')
-    section = config['task.osco-to-oscal']
-    tgt = osco_to_oscal.OscoToOscal(section)
+    config.remove_option('task.osco-result-to-oscal-ar', 'output-dir')
+    section = config['task.osco-result-to-oscal-ar']
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.simulate()
     assert retval == TaskOutcome.SIM_FAILURE
     assert len(os.listdir(str(tmp_path))) == 0
@@ -135,11 +135,11 @@ def test_osco_simulate_no_ouput_dir(tmp_path):
 def test_osco_simulate_input_fetcher(tmp_path):
     """Test simulate call OSCO fetcher json data."""
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal-fetcher.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar-fetcher.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     section['output-dir'] = str(tmp_path)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.simulate()
     assert retval == TaskOutcome.SIM_SUCCESS
     assert len(os.listdir(str(tmp_path))) == 0
@@ -148,11 +148,11 @@ def test_osco_simulate_input_fetcher(tmp_path):
 def test_osco_simulate_input_bad_yaml(tmp_path):
     """Test simulate call OSCO bad yaml data."""
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal-bad-yaml.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar-bad-yaml.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     section['output-dir'] = str(tmp_path)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.simulate()
     assert retval == TaskOutcome.SIM_FAILURE
     assert len(os.listdir(str(tmp_path))) == 0
@@ -164,13 +164,13 @@ def test_osco_execute(tmp_path, monkeypatch: MonkeyPatch):
     monkeypatch.setattr(uuid, 'uuid4', monkeybusiness.uuid_mock1)
     osco.OscoTransformer.set_timestamp('2021-02-24T19:31:13+00:00')
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     d_expected = pathlib.Path(section['output-dir'])
     d_produced = tmp_path
     section['output-dir'] = str(d_produced)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.SUCCESS
     list_dir = os.listdir(d_produced)
@@ -188,14 +188,14 @@ def test_osco_execute_checking(tmp_path, monkeypatch: MonkeyPatch):
     monkeypatch.setattr(uuid, 'uuid4', monkeybusiness.uuid_mock1)
     osco.OscoTransformer.set_timestamp('2021-02-24T19:31:13+00:00')
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     d_expected = pathlib.Path(section['output-dir'])
     d_produced = tmp_path
     section['output-dir'] = str(d_produced)
     section['checking'] = 'true'
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.SUCCESS
     list_dir = os.listdir(d_produced)
@@ -213,13 +213,13 @@ def test_osco_execute_1_3_5(tmp_path, monkeypatch: MonkeyPatch):
     monkeypatch.setattr(uuid, 'uuid4', monkeybusiness.uuid_mock1)
     osco.OscoTransformer.set_timestamp('2021-02-24T19:31:13+00:00')
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal-1.3.5.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar-1.3.5.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     d_expected = pathlib.Path(section['output-dir'])
     d_produced = tmp_path
     section['output-dir'] = str(d_produced)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.SUCCESS
     list_dir = os.listdir(d_produced)
@@ -237,14 +237,14 @@ def test_osco_execute_1_3_5_checking(tmp_path, monkeypatch: MonkeyPatch):
     monkeypatch.setattr(uuid, 'uuid4', monkeybusiness.uuid_mock1)
     osco.OscoTransformer.set_timestamp('2021-02-24T19:31:13+00:00')
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal-1.3.5.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar-1.3.5.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     d_expected = pathlib.Path(section['output-dir'])
     d_produced = tmp_path
     section['output-dir'] = str(d_produced)
     section['checking'] = 'true'
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.SUCCESS
     list_dir = os.listdir(d_produced)
@@ -262,13 +262,13 @@ def test_osco_execute_compressed(tmp_path, monkeypatch: MonkeyPatch):
     monkeypatch.setattr(uuid, 'uuid4', monkeybusiness.uuid_mock1)
     osco.OscoTransformer.set_timestamp('2021-02-24T19:31:13+00:00')
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal-compressed.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar-compressed.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     d_expected = pathlib.Path(section['output-dir'])
     d_produced = tmp_path
     section['output-dir'] = str(d_produced)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.SUCCESS
     list_dir = os.listdir(d_produced)
@@ -282,7 +282,7 @@ def test_osco_execute_compressed(tmp_path, monkeypatch: MonkeyPatch):
 
 def test_osco_execute_no_config(tmp_path):
     """Test execute no config call."""
-    tgt = osco_to_oscal.OscoToOscal(None)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(None)
     retval = tgt.execute()
     assert retval == TaskOutcome.FAILURE
     assert len(os.listdir(str(tmp_path))) == 0
@@ -300,13 +300,13 @@ def execute_no_overwrite_part1(tmp_path, monkeypatch: MonkeyPatch):
     monkeypatch.setattr(uuid, 'uuid4', monkeybusiness.uuid_mock1)
     osco.OscoTransformer.set_timestamp('2021-02-24T19:31:13+00:00')
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     d_expected = pathlib.Path(section['output-dir'])
     d_produced = tmp_path
     section['output-dir'] = str(d_produced)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.SUCCESS
     list_dir = os.listdir(d_produced)
@@ -324,12 +324,12 @@ def execute_no_overwrite_part2(tmp_path, monkeypatch: MonkeyPatch):
     monkeypatch.setattr(uuid, 'uuid4', monkeybusiness.uuid_mock2)
     osco.OscoTransformer.set_timestamp('2021-02-24T19:31:13+00:00')
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     section['output-overwrite'] = 'false'
     section['output-dir'] = str(tmp_path)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.FAILURE
 
@@ -337,12 +337,12 @@ def execute_no_overwrite_part2(tmp_path, monkeypatch: MonkeyPatch):
 def test_osco_execute_no_input_dir(tmp_path):
     """Test execute with no input dir call."""
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar.config')
     config.read(config_path)
-    config.remove_option('task.osco-to-oscal', 'input-dir')
-    section = config['task.osco-to-oscal']
+    config.remove_option('task.osco-result-to-oscal-ar', 'input-dir')
+    section = config['task.osco-result-to-oscal-ar']
     section['output-dir'] = str(tmp_path)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.FAILURE
     assert len(os.listdir(str(tmp_path))) == 0
@@ -351,11 +351,11 @@ def test_osco_execute_no_input_dir(tmp_path):
 def test_osco_execute_no_ouput_dir(tmp_path):
     """Test execute with no output dir call."""
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar.config')
     config.read(config_path)
-    config.remove_option('task.osco-to-oscal', 'output-dir')
-    section = config['task.osco-to-oscal']
-    tgt = osco_to_oscal.OscoToOscal(section)
+    config.remove_option('task.osco-result-to-oscal-ar', 'output-dir')
+    section = config['task.osco-result-to-oscal-ar']
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.FAILURE
     assert len(os.listdir(str(tmp_path))) == 0
@@ -364,11 +364,11 @@ def test_osco_execute_no_ouput_dir(tmp_path):
 def test_osco_execute_bad_timestamp(tmp_path):
     """Test execute with bad timestamp."""
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     section['timestamp'] = str('bogus')
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.FAILURE
     assert len(os.listdir(str(tmp_path))) == 0
@@ -380,13 +380,13 @@ def test_osco_execute_input_fetcher(tmp_path, monkeypatch: MonkeyPatch):
     monkeypatch.setattr(uuid, 'uuid4', monkeybusiness.uuid_mock1)
     osco.OscoTransformer.set_timestamp('2021-02-24T19:31:13+00:00')
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal-fetcher.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar-fetcher.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     d_expected = pathlib.Path(section['output-dir'])
     d_produced = tmp_path
     section['output-dir'] = str(d_produced)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.SUCCESS
     list_dir = os.listdir(d_produced)
@@ -404,13 +404,13 @@ def test_osco_execute_input_xml_rhel7(tmp_path, monkeypatch: MonkeyPatch):
     monkeypatch.setattr(uuid, 'uuid4', monkeybusiness.uuid_mock1)
     osco.OscoTransformer.set_timestamp('2021-02-24T19:31:13+00:00')
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal-xml-rhel7.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar-xml-rhel7.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     d_expected = pathlib.Path(section['output-dir'])
     d_produced = tmp_path
     section['output-dir'] = str(d_produced)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.SUCCESS
     list_dir = os.listdir(d_produced)
@@ -428,13 +428,13 @@ def test_osco_execute_input_xml_ocp4(tmp_path, monkeypatch: MonkeyPatch):
     monkeypatch.setattr(uuid, 'uuid4', monkeybusiness.uuid_mock1)
     osco.OscoTransformer.set_timestamp('2021-02-24T19:31:13+00:00')
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal-xml-ocp4.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar-xml-ocp4.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     d_expected = pathlib.Path(section['output-dir'])
     d_produced = tmp_path
     section['output-dir'] = str(d_produced)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.SUCCESS
     list_dir = os.listdir(d_produced)
@@ -452,13 +452,13 @@ def test_osco_execute_input_configmaps(tmp_path, monkeypatch: MonkeyPatch):
     monkeypatch.setattr(uuid, 'uuid4', monkeybusiness.uuid_mock1)
     osco.OscoTransformer.set_timestamp('2021-02-24T19:31:13+00:00')
     config = configparser.ConfigParser()
-    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-to-oscal-configmaps.config')
+    config_path = pathlib.Path('tests/data/tasks/osco/test-osco-result-to-oscal-ar-configmaps.config')
     config.read(config_path)
-    section = config['task.osco-to-oscal']
+    section = config['task.osco-result-to-oscal-ar']
     d_expected = pathlib.Path(section['output-dir'])
     d_produced = tmp_path
     section['output-dir'] = str(d_produced)
-    tgt = osco_to_oscal.OscoToOscal(section)
+    tgt = osco_result_to_oscal_ar.OscoResultToOscalAR(section)
     retval = tgt.execute()
     assert retval == TaskOutcome.SUCCESS
     list_dir = os.listdir(d_produced)
